Create a default puzzle page component
